requirements.txt
apscheduler~=3.10
python-telegram-bot~=20.0
sqlalchemy~=2.0
aiosqlite>=0.20.0
numpy>=1.24,<2
pillow>=10.0
feedparser>=6.0
httpx>=0.25
aiohttp>=3.9
pydantic-settings[yaml]>=2.0
alembic>=1.13
